Refactor out `lovatAPI.getMatches()`
